+++ packages/cardano-services/src/util/TypeormProvider/index.ts
@@ -1,0 +1,2 @@
+export * from './TypeormProvider';
+export * from './util';