+++ packages/core/src/Provider/StakePoolProvider/types/StakePoolProvider.ts
@@ -2,15 +2,19 @@ import { Cardano, Provider } from '../../..';
 import { Paginated, PaginationArgs } from '../../types/Pagination';
 import { SortFields } from '../util';
 
-type FilterCondition = 'and' | 'or';
-type SortOrder = 'asc' | 'desc';
+export type FilterCondition = 'and' | 'or';
+export type SortOrder = 'asc' | 'desc';
 
 export type SortField = typeof SortFields[number];
 
-interface StakePoolSortOptions {
+export interface StakePoolSortOptions {
   order: SortOrder;
   field: SortField;
 }
+
+export type FilterIdentifiers = Partial<
+  Pick<Cardano.PoolParameters, 'id'> & Pick<Cardano.StakePoolMetadata, 'name' | 'ticker'>
+>;
 
 export interface MultipleChoiceSearchFilter<T> {
   /**
@@ -36,16 +40,14 @@ export interface QueryStakePoolsArgs {
     /**
      * Will return results for partial matches
      */
-    identifier?: MultipleChoiceSearchFilter<
-      Partial<Pick<Cardano.PoolParameters, 'id'> & Pick<Cardano.StakePoolMetadata, 'name' | 'ticker'>>
-    >;
+    identifier?: MultipleChoiceSearchFilter<FilterIdentifiers>;
     pledgeMet?: boolean;
     status?: Cardano.StakePoolStatus[];
   };
   /**
-   * Will fetch stake pool reward history up to 3 epochs back if not specified
+   * Used for APY metric computation. It will take 3 epochs back if not specified
    */
-  rewardsHistoryLimit?: number;
+  apyEpochsBackLimit?: number;
   /**
    * Will return all stake pools matching the query if not specified
    */
@@ -54,6 +56,7 @@ export interface QueryStakePoolsArgs {
 
 export interface StakePoolStats {
   qty: {
+    activating: number;
     active: number;
     retired: number;
     retiring: number;