+++ packages/projection-typeorm/src/entity/PoolRegistration.entity.ts
@@ -2,7 +2,8 @@
 import { BigIntColumnOptions, DeleteCascadeRelationOptions } from './util';
 import { BlockEntity } from './Block.entity';
 import { Cardano } from '@cardano-sdk/core';
-import { Column, Entity, JoinColumn, ManyToOne, PrimaryColumn } from 'typeorm';
+import { Column, Entity, JoinColumn, ManyToOne, OneToOne, PrimaryColumn } from 'typeorm';
+import { PoolMetadataEntity } from './PoolMetadata.entity';
 import { StakePoolEntity } from './StakePool.entity';
 
 @Entity()
@@ -39,6 +40,8 @@ export class PoolRegistrationEntity {
   @JoinColumn()
   @ManyToOne(() => StakePoolEntity, (stakePool) => stakePool.registrations, DeleteCascadeRelationOptions)
   stakePool?: StakePoolEntity;
+  @OneToOne(() => PoolMetadataEntity, (metadata) => metadata.poolUpdate)
+  metadata?: PoolMetadataEntity | null;
   @ManyToOne(() => BlockEntity, DeleteCascadeRelationOptions)
   @JoinColumn()
   block?: BlockEntity;