+++ packages/core/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.12.2](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/core@0.12.1...@cardano-sdk/core@0.12.2) (2023-06-01)
+
+### Features
+
+* add HandleProvider interface and handle support implementation to TxBuilder ([f209095](https://github.com/input-output-hk/cardano-js-sdk/commit/f2090952c8a0512fc589674b876f3a27be403140))
+
 ## [0.12.1](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/core@0.12.0...@cardano-sdk/core@0.12.1) (2023-05-24)
 
 **Note:** Version bump only for package @cardano-sdk/core