+++ packages/projection-typeorm/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.3](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/projection-typeorm@0.2.2...@cardano-sdk/projection-typeorm@0.2.3) (2023-06-01)
+
+### Features
+
+* **projection-typeorm:** add stake pools metrics computation job schedule ([a5f56e9](https://github.com/input-output-hk/cardano-js-sdk/commit/a5f56e9105ae2ed1fddb2a64c1c6aeef5016f51a))
+
 ## [0.2.2](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/projection-typeorm@0.2.1...@cardano-sdk/projection-typeorm@0.2.2) (2023-05-24)
 
 **Note:** Version bump only for package @cardano-sdk/projection-typeorm