+++ packages/projection/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.3](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/projection@0.6.2...@cardano-sdk/projection@0.6.3) (2023-06-01)
+
+**Note:** Version bump only for package @cardano-sdk/projection
+
 ## [0.6.2](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/projection@0.6.1...@cardano-sdk/projection@0.6.2) (2023-05-24)
 
 **Note:** Version bump only for package @cardano-sdk/projection